requirements.txt
botbuilder-integration-aiohttp>=4.15.0
aiohttp
botbuilder-core
botbuilder-schema
openai